services/api-service/pyproject.toml
[project]
name = "api-service"
version = "0.1.0"
description = "Copycat API Gateway - Central API for managing and monitoring Copycat services"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.119.1",
    "uvicorn[standard]==0.34.0",
    "pydantic==2.10.6",
    "pydantic-settings==2.11.0",
    "google-cloud-firestore==2.21.0",
    "google-cloud-bigquery==3.38.0",
    "google-auth==2.42.0",
    "google-auth-httplib2==0.2.0",
    "httpx==0.28.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==1.2.0",
    "pytest-cov==7.0.0",
    "ruff==0.8.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
